STARCH-300 selectively disable rivet loading and/or error dialog event listeners from environment config
